WASD selection pt. 1
There are now selected (red) versions of the each piece along with dark and light, all representing which color square the piece is rested upon. (Example: A dark black bishop would be a black bishop on a dark square) There is also a function to select (change a piece/square to it's corresponding red square) a piece, however there it is not being used.