Function to check for checkmate pt. 1
There is now a function to identify checkmate, which sees if either black or white is in check and simulates all moves to see if any prevent check, if not, it is checkmate. This function is still in progress and does not work, since it causes bugs by manually simulating moves on the board (as a placeholder).