Function to check for checkmate pt. 2
It now has a function inside of it to see if it is a check for a simulated board, and it runs this check for every possible move the color in check can make, trying to see if any move results in it not being a check, otherwise it is checkmate. This function is still buggy however and does not work (yet).